+++ bricks/ai/src/chat-view/ChatViewContext.ts
@@ -1,5 +1,6 @@
 import { createContext, useContext } from "react";
 import { AgentDetailItem } from "./components/QuickAnswerList/index.js";
+import { SessionItem } from "./ChatService.js";
 
 export type Role = "guide" | "user" | "assistant";
 
@@ -11,17 +12,16 @@ export interface MessageItemContent {
 
 export interface MessageItem {
   role: Role;
-  content: MessageItemContent[];
-  ctime?: number | string;
-  key?: number;
+  content: MessageItemContent;
+  created?: number | string;
+  key?: number | string;
+  conversationId?: string;
+  taskId?: string;
+  agentId?: string;
   chatting?: boolean;
-}
-
-export interface SessionItem {
-  title: string;
-  ctime: number;
-  id: string;
-  active?: boolean;
+  tag?: {
+    isLike?: boolean;
+  };
 }
 
 export interface QuickAnswerConfig {
@@ -30,8 +30,12 @@ export interface QuickAnswerConfig {
 }
 
 interface ChatViewContextProps {
-  activeSessionId: string;
+  sessionEnd: boolean;
+  sessionLoading: boolean;
+  activeSessionId?: string;
   sessionList: SessionItem[];
+  msgEnd: boolean;
+  msgLoading: boolean;
   msgList: MessageItem[];
   msgItem?: MessageItem;
   loading: boolean;
@@ -39,11 +43,14 @@ interface ChatViewContextProps {
   searchStr: string;
   showLike: boolean;
   quickAnswerConfig?: QuickAnswerConfig;
+  handleIsLike: (id: string, isLike: boolean) => Promise<boolean>;
   handleChat: (str: string) => void;
   stopChat: () => void;
-  updateSession: (id: string) => void;
+  checkSession: (id?: string, isInit?: boolean) => void;
   createSession: () => void;
+  deleteSession: (ids: string[]) => Promise<boolean>;
   setSearchStr: (str: string) => void;
+  querySessionHistory: (limit?: number) => void;
 }
 
 export const ChatViewContext = createContext<ChatViewContextProps>(