+++ bricks/ai/src/chat-view/index.tsx
@@ -11,6 +11,7 @@ import loadStyleText from "./components/loading.shadow.css";
 import markdownStyleText from "./components/MessageItem/markdown.shadow.css";
 import quickAnswerListStyleText from "./components/QuickAnswerList/quickAnswerList.shadow.css";
 import "@next-core/theme";
+import "./host-context.css";
 
 const { defineElement, property } = createDecorators();
 
@@ -114,18 +115,25 @@ export function ChatViewComponent({
   enterInterval,
 }: ChatViewProps) {
   const {
+    sessionEnd,
+    sessionLoading,
     activeSessionId,
+    msgEnd,
+    msgLoading,
     msgList,
     msgItem,
     sessionList,
     loading,
     chatting,
     searchStr,
+    handleIsLike,
     handleChat,
     stopChat,
     createSession,
-    updateSession,
+    deleteSession,
+    checkSession,
     setSearchStr,
+    querySessionHistory,
   } = useChatViewInfo({
     agentId,
     enterInterval,
@@ -134,8 +142,12 @@ export function ChatViewComponent({
   return (
     <ChatViewContext.Provider
       value={{
+        sessionEnd,
+        sessionLoading,
         activeSessionId,
         sessionList,
+        msgEnd,
+        msgLoading,
         msgList,
         msgItem,
         chatting,
@@ -143,11 +155,14 @@ export function ChatViewComponent({
         searchStr,
         showLike,
         quickAnswerConfig,
+        handleIsLike,
         handleChat,
         stopChat,
         createSession,
-        updateSession,
+        deleteSession,
+        checkSession,
         setSearchStr,
+        querySessionHistory,
       }}
     >
       <div className="chat-view-container">