+++ bricks/presentational/src/bootstrap.ts
@@ -8,3 +8,5 @@ import "./info-card-item/index.js";
 import "./humanize-time/index.js";
 import "./statistics-card/index.js";
 import "./loading-step/index.js";
+import "./current-time/index.js";
+import "./carousel-text/index.js";