+++ bricks/api-market/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.13](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/api-market@0.1.12...@next-bricks/api-market@0.1.13) (2024-04-16)
+
+**Note:** Version bump only for package @next-bricks/api-market
+
+
+
+
+
 ## [0.1.12](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/api-market@0.1.11...@next-bricks/api-market@0.1.12) (2024-04-08)
 
 **Note:** Version bump only for package @next-bricks/api-market