+++ bricks/form-renderer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.51](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form-renderer@0.1.50...@next-bricks/form-renderer@0.1.51) (2024-04-16)
+
+**Note:** Version bump only for package @next-bricks/form-renderer
+
+
+
+
+
 ## [0.1.50](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form-renderer@0.1.49...@next-bricks/form-renderer@0.1.50) (2024-04-08)
 
 **Note:** Version bump only for package @next-bricks/form-renderer