+++ bricks/visual-builder/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.8](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/visual-builder@1.7.7...@next-bricks/visual-builder@1.7.8) (2024-04-16)
+
+**Note:** Version bump only for package @next-bricks/visual-builder
+
+
+
+
+
 ## [1.7.7](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/visual-builder@1.7.6...@next-bricks/visual-builder@1.7.7) (2024-04-08)
 
 **Note:** Version bump only for package @next-bricks/visual-builder