+++ shared/common-bricks/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.41.0](https://github.com/easyops-cn/next-bricks/compare/@next-shared/common-bricks@0.40.0...@next-shared/common-bricks@0.41.0) (2024-04-17)
+
+
+### Features
+
+* feat carousel text ([b849cc9](https://github.com/easyops-cn/next-bricks/commit/b849cc966b5aec9439088d9c2cc4853ef4e22e66))
+
+
+
+
+
+# [0.40.0](https://github.com/easyops-cn/next-bricks/compare/@next-shared/common-bricks@0.39.0...@next-shared/common-bricks@0.40.0) (2024-04-16)
+
+
+### Features
+
+* feat current time bricks ([5181177](https://github.com/easyops-cn/next-bricks/commit/518117702e7284fdc27ca8d78daea82e299dc53b))
+
+
+
+
+
 # [0.39.0](https://github.com/easyops-cn/next-bricks/compare/@next-shared/common-bricks@0.38.0...@next-shared/common-bricks@0.39.0) (2024-03-29)