+++ shared/general/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.20](https://github.com/easyops-cn/next-bricks/compare/@next-shared/general@0.8.19...@next-shared/general@0.8.20) (2024-04-16)
+
+**Note:** Version bump only for package @next-shared/general
+
+
+
+
+
 ## [0.8.19](https://github.com/easyops-cn/next-bricks/compare/@next-shared/general@0.8.18...@next-shared/general@0.8.19) (2024-04-08)
 
 **Note:** Version bump only for package @next-shared/general